[FIX] account_financial_report: Ability to filter by code
The way accounts are collected has been modified, as they were returning ‘NewId’, which meant that filtering was not possible at any point. I check whether the first position is already a ‘NewId’ and, if so, I browse the IDs so that ‘self.account_ids’ is filled in correctly and filtered correctly.